feat(print): order export/normal

+++ apps/server/src/aorder/aorder.service.ts
@@ -1,17 +1,19 @@
+import { Prisma } from '@prisma/client';
+import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { AOrderRepository } from './aorder.repository';
 import { CreateAOrderDto } from './dto/create-aorder.dto';
 import { UpdateAOrderDto } from './dto/update-aorder.dto';
 import { AOrderDiscrimination } from './types/aorder-discrimination.enum';
-import { Prisma } from '@prisma/client';
 import { FindManyDto } from './dto/find-many.dto';
 import { UpdateManyAOrderDto } from './dto/update-many-aorder.dto';
 import { AOrderProcess } from './aorder.process';
 import { PrintService } from '../print/print.service';
 import { FileService } from '../file/file.service';
-import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { AOrderProductProcess } from './aorder-product.process';
 import { AOrderPayload } from './types/aorder-payload';
 import { ContactService } from '../contact/contact.service';
+import { AOrderPayloadRelation } from './types/aorder-payload-relation';
+
 type CommonAOrderDto = Partial<Omit<CreateAOrderDto, 'pickup' | 'repair'>>;
 type CommonAOrderInput = Partial<Omit<Prisma.aorderCreateInput, 'pickup' | 'repair'>>;
 
@@ -54,7 +56,7 @@ export class AOrderService {
             id: true,
             name: true,
             color: true,
-          }
+          },
         },
         ...(this.type === AOrderDiscrimination.PURCHASE && {
           pickup: {
@@ -76,16 +78,18 @@ export class AOrderService {
         ...(this.type && { discr: this.type }),
         ...(search && { order_nr: { contains: search } }),
         ...(status && { status_id: { equals: status } }),
-        ...this.getPartnerWhereInput({createdBy, partner, email}),
+        ...this.getPartnerWhereInput({ createdBy, partner, email }),
       },
       orderBy: Object.keys(query?.orderBy || {})?.length ? query.orderBy : { id: 'desc' },
     };
 
     const result = await this.repository.findAll(params);
-    
+
     return {
       count: result.count,
-      data: result.data.map(order => new AOrderProductProcess(new AOrderProcess(order).run()).run()),
+      data: result.data.map(
+        (order) => new AOrderProductProcess(new AOrderProcess(order).run()).run(),
+      ),
     };
   }
 
@@ -93,16 +97,16 @@ export class AOrderService {
     const params: Prisma.aorderFindUniqueArgs = {
       where: {
         id,
-        ...this.getPartnerWhereInput({email}),
+        ...this.getPartnerWhereInput({ email }),
       },
     };
 
-    const order = <AOrderPayload>await this.repository.findOne(this.commonIncludePart(params));
+    const order = <AOrderPayload> await this.repository.findOne(this.commonIncludePart(params));
 
     if (!order || (this.type && order?.discr !== this.type)) {
       throw new NotFoundException(`Order with ID ${id} not found`);
     }
-    
+
     return new AOrderProcess(order).run();
   }
 
@@ -116,27 +120,27 @@ export class AOrderService {
     const params: Prisma.aorderCreateArgs = {
       data: {
         ...await this.processCreateOrUpdateOrderInput(commonDto),
-        order_nr: commonDto.order_nr || 'TEMP' + Math.floor(Date.now() / 1000).toString(),
+        order_nr: commonDto.order_nr || `TEMP${Math.floor(Date.now() / 1000).toString()}`,
         discr: this.type,
         order_date: new Date(),
         ...(pickup && { pickup: { create: { ...pickup } } }),
-        ...(repair && { repair: { create: { ...repair } } })
-      }
+        ...(repair && { repair: { create: { ...repair } } }),
+      },
     };
 
-    const order = <AOrderPayload>await this.repository.create(this.commonIncludePart(params));
+    const order = <AOrderPayload> await this.repository.create(this.commonIncludePart(params));
 
     if (commonDto.order_nr === undefined) {
-      const { id, order_date } = order;
+      const { id, order_date: orderDate } = order;
 
-      const order_nr = order_date.getFullYear() + id.toString().padStart(6, "0");
+      const orderNumber = orderDate.getFullYear() + id.toString().padStart(6, '0');
 
       try {
         await this.repository.update({
           where: { id },
-          data: { order_nr },
+          data: { order_nr: orderNumber },
         });
-        order.order_nr = order_nr;
+        order.order_nr = orderNumber;
       } catch (e) {
         this.repository.deleteMany([id]);
         throw e;
@@ -156,11 +160,11 @@ export class AOrderService {
     };
 
     const params: Prisma.aorderUpdateArgs = {
-      data: data,
-      where: { id }
+      data,
+      where: { id },
     };
 
-    const order = <AOrderPayload>await this.repository.update(this.commonIncludePart(params));
+    const order = <AOrderPayload> await this.repository.update(this.commonIncludePart(params));
 
     return new AOrderProcess(order).run();
   }
@@ -169,61 +173,71 @@ export class AOrderService {
     return this.repository.updateMany({
       data: updateManyOrderDto.order,
       where: {
-        id: { in: updateManyOrderDto.ids }
-      }
+        id: { in: updateManyOrderDto.ids },
+      },
     });
   }
 
   async deleteOne(id: number) {
-    const order = await this.findOne(id);
+    const order = <AOrderPayloadRelation> await this.findOne(id);
     if (order.discr === AOrderDiscrimination.PURCHASE) {
-      order?.pickup?.['afile']?.length &&
+      if (order?.pickup?.afile?.length) {
         await this.fileService.deleteMany(
-          order?.pickup?.['afile']?.map((file) => file.id)
+          order?.pickup?.afile?.map((file: { id: unknown; }) => file.id),
         );
-      order?.pickup?.id &&
+      }
+      if (order?.pickup?.id) {
         await this.repository.deletePickup(order?.pickup?.id);
+      }
     } else if (order.discr === AOrderDiscrimination.SALE) {
-      order['afile']?.length &&
-        this.fileService.deleteMany(order['afile']?.map((file) => file.id));
+      if (order.afile?.length) {
+        this.fileService.deleteMany(order.afile?.map((file) => file.id));
+      }
     }
 
     return this.repository.deleteOne(id);
   }
-  
+
   async deleteFiles(id: number, fileIds: number[]) {
     await this.findOne(id);
-    
+
     return this.fileService.deleteMany(fileIds);
   }
 
-  async findByIds(ids: number[], email?: string) {
+  async findByIds(ids: number[], product?: boolean, email?: string) {
     const params: Prisma.aorderFindManyArgs = {
       where: {
         id: { in: ids },
-        ...this.getPartnerWhereInput({email}),
+        ...this.getPartnerWhereInput({ email }),
       },
       orderBy: {
         id: 'asc',
       },
     };
 
-    const result = <AOrderPayload[]>await this.repository.findBy(this.commonIncludePart(params));
+    const result = <AOrderPayload[]> await this.repository
+      .findBy(product ? this.productIncludePart(params) : this.commonIncludePart(params));
 
-    return result.map(order => new AOrderProcess(order).run());
+    return result.map((order) => new AOrderProcess(order).run());
   }
 
   async printAOrders(ids: number[], email?: string): Promise<Buffer> {
-    const aorders = await this.findByIds(ids, email);
+    const aorders = await this.findByIds(ids, false, email);
     return this.printService.printAOrders(aorders);
   }
 
-  protected async processCreateOrUpdateOrderInput(orderDto: CommonAOrderDto): Promise<CommonAOrderInput> {
+  async printExport(ids: number[], email?: string): Promise<Buffer> {
+    const aorders = await this.findByIds(ids, true, email);
+    return this.printService.printExport(aorders);
+  }
+
+  protected async processCreateOrUpdateOrderInput(orderDto: CommonAOrderDto):
+  Promise<CommonAOrderInput> {
     const {
-      status_id,
-      customer_id,
+      status_id: statusIdDto,
+      customer_id: customerIdDto,
       customer,
-      supplier_id,
+      supplier_id: supplierIdDto,
       supplier,
       ...rest
     } = orderDto;
@@ -231,7 +245,7 @@ export class AOrderService {
     let customerId: number;
     let supplierId: number;
 
-    if (!customer_id && (customer?.company_id || customer?.company_name)) {
+    if (!customerIdDto && (customer?.company_id || customer?.company_name)) {
       customerId = (await this.contactService.checkExists({
         ...customer,
         company_is_partner: false,
@@ -239,7 +253,7 @@ export class AOrderService {
         company_is_supplier: false,
       }))?.id;
     }
-    if (!supplier_id && (supplier?.company_id || supplier?.company_name)) {
+    if (!supplierIdDto && (supplier?.company_id || supplier?.company_name)) {
       supplierId = (await this.contactService.checkExists({
         ...supplier,
         company_is_partner: false,
@@ -250,17 +264,20 @@ export class AOrderService {
 
     const data: CommonAOrderInput = {
       ...rest,
-      ...(status_id && { order_status: { connect: { id: status_id } } }),
-      ...((customer_id || customerId) && { contact_aorder_customer_idTocontact: { connect: { id: customer_id || customerId } } }),
-      ...((supplier_id || supplierId) && { contact_aorder_supplier_idTocontact: { connect: { id: supplier_id || supplierId } } }),
+      ...(statusIdDto && { order_status: { connect: { id: statusIdDto } } }),
+      ...((customerIdDto || customerId)
+      && { contact_aorder_customer_idTocontact: { connect: { id: customerIdDto || customerId } } }),
+      ...((supplierIdDto || supplierId)
+      && { contact_aorder_supplier_idTocontact: { connect: { id: supplierIdDto || supplierId } } }),
     };
 
     return data;
   }
 
   protected commonIncludePart<T extends Prisma.aorderDefaultArgs>(params: T): T {
-    params.include = {
-      ...params.include,
+    let { include } = params;
+    include = {
+      ...include,
       order_status: {
         select: {
           id: true,
@@ -269,11 +286,7 @@ export class AOrderService {
         },
       },
       product_order: {
-        select: {
-          id: true,
-          product_id: true,
-          price: true,
-          quantity: true,
+        include: {
           product: {
             select: {
               sku: true,
@@ -288,16 +301,16 @@ export class AOrderService {
           aservice: {
             select: {
               status: true,
-              price: true
-            }
-          }
+              price: true,
+            },
+          },
         },
       },
     };
 
     if (this.type !== AOrderDiscrimination.SALE) {
-      params.include = {
-        ...params.include,
+      include = {
+        ...include,
         pickup: {
           include: {
             afile: {
@@ -311,14 +324,14 @@ export class AOrderService {
           },
         },
         contact_aorder_supplier_idTocontact: {
-          select: this.getContactSelect()
+          select: this.getContactSelect(),
         },
       };
     }
-    
+
     if (this.type !== AOrderDiscrimination.PURCHASE) {
-      params.include = {
-        ...params.include,
+      include = {
+        ...include,
         afile: {
           select: {
             id: true,
@@ -328,13 +341,45 @@ export class AOrderService {
           },
         },
         contact_aorder_customer_idTocontact: {
-          select: this.getContactSelect()
+          select: this.getContactSelect(),
         },
-        repair: true
+        repair: true,
       };
     }
 
-    return params;
+    return {
+      ...params,
+      include,
+    };
+  }
+
+  protected productIncludePart<T extends Prisma.aorderDefaultArgs>(params: T): T {
+    let { include } = this.commonIncludePart(params);
+    include = {
+      ...include,
+      product_order: {
+        include: {
+          product: {
+            include: {
+              product_attribute_product_attribute_product_idToproduct: {
+                include: {
+                  attribute: {
+                    include: {
+                      attribute_option: true,
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    };
+
+    return {
+      ...params,
+      include,
+    };
   }
 
   protected getContactSelect(): Prisma.contactSelect {
@@ -360,7 +405,7 @@ export class AOrderService {
               },
             },
           },
-        }
+        },
       },
       name: true,
       email: true,
@@ -383,8 +428,14 @@ export class AOrderService {
     return {
       ...((createdBy || partner || email) && {
         OR: [
-          { contact_aorder_supplier_idTocontact: this.getContactWhereInput({createdBy, partner, email}) },
-          { contact_aorder_customer_idTocontact: this.getContactWhereInput({createdBy, partner, email}) },
+          {
+            contact_aorder_supplier_idTocontact: this
+              .getContactWhereInput({ createdBy, partner, email }),
+          },
+          {
+            contact_aorder_customer_idTocontact: this
+              .getContactWhereInput({ createdBy, partner, email }),
+          },
         ],
       }),
     };
@@ -402,7 +453,11 @@ export class AOrderService {
       ...(email && {
         OR: [
           { email },
-          { company_contact_company_idTocompany: { company: { companyContacts: { some: { email } } } } },
+          {
+            company_contact_company_idTocompany: {
+              company: { companyContacts: { some: { email } } },
+            },
+          },
         ],
       }),
     };
